// src/app/models/information.ts
export interface Information {
    userId: number;
    totalSales: number;

    selledAlbums: number;
    selledDonations: number;
    activeSubscriptions: number;

    albumsAmount: number;
    donationsAmount: number;
    subscriptionsAmount: number;
    chatAmount: number;
    totalAmount: number;

    publishedAlbums: number;
    photos: number;
    videos: number;
}